refactor: folders restructure
