+++ src/App.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import './App.css';
+import './styles/App.css';
 import Hero from './components/Hero.jsx';
 import Features from "./components/Features.jsx";
 import ContactForm2 from './components/ContactForm2.jsx';
